Reify client as an explicit argument to all reporting procedures
This in the main change I wanted to make to cognician.dogstatsd from
which we forked.

The reason for this change is that storing configuration in an atom
which was then implicitly dereferenced does not work well with
component systems, such as integrant.

This breaking change makes it possible to easily wrap this library in
an integrant component. Without this you need to resort to hacks, such
as configuring and initialising this component outside of the
integrant system, or forcing it to be artificially earlier than any of
its dependencies in the integrant dependency graph.